requirements.txt
Django>=5.0
psycopg2-binary
pillow==12.1.0
django-cleanup==9.0.0
djangorestframework==3.16.1
requests==2.32.5
PyJWT==2.10.1
cryptography==46.0.3
drf-spectacular==0.29.0